Rework restrict code into TeamMatchModule
