+++ core/src/main/java/tc/oc/pgm/api/Modules.java
@@ -82,7 +82,6 @@ import tc.oc.pgm.modules.PlayableRegionMatchModule;
 import tc.oc.pgm.modules.PlayableRegionModule;
 import tc.oc.pgm.modules.SoundsMatchModule;
 import tc.oc.pgm.modules.StatsMatchModule;
-import tc.oc.pgm.modules.TeamRestrictMatchModule;
 import tc.oc.pgm.modules.TimeLockModule;
 import tc.oc.pgm.modules.ToolRepairMatchModule;
 import tc.oc.pgm.modules.ToolRepairModule;
@@ -160,7 +159,6 @@ public interface Modules {
     register(ObserverToolsMatchModule.class, new ObserverToolsMatchModule.Factory());
     register(FireworkMatchModule.class, FireworkMatchModule::new);
     register(StatsMatchModule.class, StatsMatchModule::new);
-    register(TeamRestrictMatchModule.class, TeamRestrictMatchModule::new);
 
     // MatchModules that require other dependencies
     register(GoalMatchModule.class, new GoalMatchModule.Factory());
